server/requirements.txt
requests
gevent-socketio
flask
